+++ packages/moonstone/IncrementSlider/IncrementSlider.js
@@ -151,6 +151,14 @@ const IncrementSliderBase = kind({
 		disabled: PropTypes.bool,
 
 		/**
+		 * Indicates that the slider knob is being dragged.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		dragging: PropTypes.bool,
+
+		/**
 		 * Shows the tooltip, when present.
 		 * @type {Boolean}
 		 * @public
@@ -471,6 +479,7 @@ const IncrementSliderBase = kind({
 		decrementDisabled,
 		decrementIcon,
 		disabled,
+		dragging,
 		focused,
 		id,
 		incrementAriaLabel,
@@ -514,7 +523,7 @@ const IncrementSliderBase = kind({
 					disabled={decrementDisabled}
 					onTap={onDecrement}
 					onSpotlightDisappear={onDecrementSpotlightDisappear}
-					spotlightDisabled={spotlightDisabled}
+					spotlightDisabled={spotlightDisabled || dragging}
 				>
 					{decrementIcon}
 				</IncrementSliderButton>
@@ -551,7 +560,7 @@ const IncrementSliderBase = kind({
 					disabled={incrementDisabled}
 					onTap={onIncrement}
 					onSpotlightDisappear={onIncrementSpotlightDisappear}
-					spotlightDisabled={spotlightDisabled}
+					spotlightDisabled={spotlightDisabled || dragging}
 				>
 					{incrementIcon}
 				</IncrementSliderButton>