+++ packages/moonstone/Slider/Slider.js
@@ -86,6 +86,14 @@ const SliderBase = kind({
 		css: PropTypes.object,
 
 		/**
+		 * Indicates that the slider knob is being dragged.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		dragging: PropTypes.bool,
+
+		/**
 		 * Indicates that the slider has gained focus and if the tooltip is present, it will be
 		 * shown.
 		 *
@@ -259,6 +267,7 @@ const SliderBase = kind({
 	render: ({css, focused, tooltip, ...rest}) => {
 		delete rest.activateOnFocus;
 		delete rest.active;
+		delete rest.dragging;
 		delete rest.knobStep;
 		delete rest.onActivate;